requirements.txt
google.cloud
google-cloud-speech
pydub
SpeechRecognition
spleeter